[DEV] Properly reset section on creation
Before this commit, the consumer could access the history of the
section even in the case when the producer had re-created the section
(and the consumer had re-open it).

From this commit, on section (re)creation, the whole section is filled
with invalid samples, and the consumer can only search through the
valid samples.

Change-Id: I3dd4fe00bfa08a97d361afb861041685f7529464
Reviewed-on: https://gerrit.parrot.biz/53821
Reviewed-by: Julien Hervé <julien.herve@parrot.com>